+++ Assets/ProceduralStages/Terrain/TerrainType.cs
@@ -8,9 +8,27 @@ namespace ProceduralStages
 {
     public enum TerrainType
     {
-        None,
+        Random,
         OpenCaves,
         Islands,
         TunnelCaves
+    }
+
+    public static class TerrainTypeExtensions
+    {
+        public static string GetName(this TerrainType terrainType)
+        {
+            switch (terrainType)
+            {
+                case TerrainType.OpenCaves:
+                    return "Open caves";
+                case TerrainType.TunnelCaves:
+                    return "Tunnel caves";
+                case TerrainType.Islands:
+                    return "Islands";
+                default:
+                    return "?";
+            }
+        }
     }
 }